Set up tRPC with Express, including a Hello Example
trpc.ts
Defined a basic tRPC server setup with a public API endpoint.

index.ts
This demonstrates a basic example of building a simple `API endpoint` that returns "hello" with tRPC server.

client.ts
It acts as a bridge between your React application and the tRPC server.

Providers.tsx
This sets up providers for using tRPC and React Query in a React application, enabling communication with a tRPC server and managing data fetching and caching on the client-side.

server.ts
creating a middleware ,if we get a  request to this endpoint `/api/trpc` we want to forward it to `appRouter`